fix bug si erreur de saisie

+++ src/model/Command.php
@@ -15,14 +15,18 @@ class Command
     public function detail($line): void {
         $lineVerif = preg_match('/detail/', $line, $matches);
         $contactId = trim(str_replace($matches, '',$line));
-        $contactRepository = new ContactManager();
-        $contactRepository->connection = new DBConnect();
-        $contact = $contactRepository->findById($contactId);
-        if($contact->getName() === null){
-            echo "Aucun contact avec cet identifiant, saisissez la commande \"list\" pour voir les identifiants disponibles \n";
+        if (is_numeric($contactId)){
+            $contactRepository = new ContactManager();
+            $contactRepository->connection = new DBConnect();
+            $contact = $contactRepository->findById($contactId);
+            if($contact->getName() === null){
+                echo "Aucun contact avec cet identifiant, saisissez la commande \"list\" pour voir les identifiants disponibles \n";
+            } else {
+                $contact->toStringDetail();
+            }
         } else {
-            $contact->toStringDetail();
-        }
+        echo "L'identifiant doit être un chiffre.\n";          
+        } 
     }
 
     public function create(): void
@@ -33,27 +37,39 @@ class Command
         $valid = null;
         $rawCreateContact = readline("Saisissez le nom, l'adresse e-mail et le numéro de téléphone, séparés par des virgules : ");
         $createContact = [];
-        $createContact = explode(',', $rawCreateContact, 3);
-        $name = trim($createContact[0]);
-        $email = trim($createContact[1]);
-        $phone = trim($createContact[2]);
-        if($name ===''){
-                echo "! ERREUR ! Le nom ne doit comporter que des lettres\n";
-            } elseif (!filter_var($email, FILTER_VALIDATE_EMAIL))  {
-                echo "! ERREUR ! Saissez une adresse e-mail valide\n";
-            } elseif (strlen($phone) < 10){
-                echo "! ERREUR ! Saissez un numéro de téléphone valide\n";
+        if (substr_count($rawCreateContact, ',', 0) === 2){
+            $createContact = explode(',', $rawCreateContact, 3);
+            $name = trim($createContact[0]);
+            $email = trim($createContact[1]);
+            $phone = trim($createContact[2]);
+            if (
+                empty($name)||
+                empty($email)||
+                empty($phone)||
+                is_null($name)||
+                is_null($email)||
+                is_null($phone)){
+                echo "Tous les champs sont obligatoire pour créer un contact.\n";
+                } elseif (preg_match('/\d/i', $name)){
+                    echo "! ERREUR ! Le nom ne doit comporter que des lettres\n";
+                } elseif (!filter_var($email, FILTER_VALIDATE_EMAIL))  {
+                    echo "! ERREUR ! Saissez une adresse e-mail valide\n";
+                } elseif (strlen($phone) < 10){
+                    echo "! ERREUR ! Saissez un numéro de téléphone valide\n";
+                }
+            else {
+                $valid = readline("Vous avez saisi : $name, $email, $phone, saisissez \"oui\" pour confirmer votre commande : ");
             }
-        else {
-            $valid = readline("Vous avez saisi : $name, $email, $phone, saisissez \"oui\" pour confirmer votre commande : ");
-        }
-        if($valid === "oui"){
-            $createRepository = new ContactManager();
-            $createRepository->connection = new DBConnect();
-            $createContact = $createRepository->createContact($name, $email, $phone);
-        }
-        else{
-            echo "Ressaisissez votre contact en tapant la commande \"create\"\n";
+            if($valid === "oui"){
+                $createRepository = new ContactManager();
+                $createRepository->connection = new DBConnect();
+                $createContact = $createRepository->createContact($name, $email, $phone);
+            }
+            else{
+                echo "Ressaisissez votre contact en tapant la commande \"create\".\n";
+            }
+        } else {
+            echo "Les données saisies ne sont pas conformes, veuillez refaire votre saisie.\n";
         }
     }
 
@@ -78,24 +94,27 @@ class Command
                 $phone = null;
                 $updValid = null;
                 $updContact = [];
-                $updContact = explode(',', $rawUpdContact, 3);
-                if(trim($updContact[0] === '-')){$name = $oldName;} else {$name = trim($updContact[0]);}
-                if(trim($updContact[1] === '-')){$email = $oldEmail;} else {$email = trim($updContact[1]);}
-                if(trim($updContact[2] === '-')){$phone = $oldPhone;} else {$phone = trim($updContact[2]);}
-
-                if(preg_match('/\d/i', $name) && $name !== null){
-                echo "! ERREUR ! Le nom ne doit comporter que des lettres : \n";
-                } elseif (!filter_var($email, FILTER_VALIDATE_EMAIL) && $email !== null)  {
-                    echo "! ERREUR ! Saissez une adresse e-mail valide, \n";
-                } elseif (strlen($phone) < 10 && $phone !== null){
-                echo "! ERREUR ! Saissez un numéro de téléphone valide, \n";
+                if (substr_count($rawUpdContact, ',', 0) === 2){
+                    $updContact = explode(',', $rawUpdContact, 3);
+                    if(trim($updContact[0] === '-')){$name = $oldName;} else {$name = trim($updContact[0]);}
+                    if(trim($updContact[1] === '-')){$email = $oldEmail;} else {$email = trim($updContact[1]);}
+                    if(trim($updContact[2] === '-')){$phone = $oldPhone;} else {$phone = trim($updContact[2]);}                
+                    if(preg_match('/\d/i', $name) && $name !== null || $name ==='' || empty($name)){
+                    echo "! ERREUR ! Le nom ne doit comporter que des lettres : \n";
+                    } elseif (!filter_var($email, FILTER_VALIDATE_EMAIL) && $email !== null)  {
+                        echo "! ERREUR ! Saissez une adresse e-mail valide, \n";
+                    } elseif (strlen($phone) < 10 && $phone !== null){
+                    echo "! ERREUR ! Saissez un numéro de téléphone valide, \n";
+                    } else {
+                    $updValid = readline("Vous avez saisi : $name, $email, $phone.\nSi vous souhaitez mettre à jour le contact avec ces données, saisissez \"oui\" :");
+                    }
+                    if ($updValid === "oui"){
+                        $updateContact = $updateRepository->updateContact($contactId, $name, $email, $phone);
+                    } else {
+                    echo "La modification n'a pas eu lieu. Vous pouvez recommencer votre modification en saisissant la commande \"modify\".\n";
+                    }
                 } else {
-                $updValid = readline("Vous avez saisi : $name, $email, $phone.\n Si vous souhaitez mettre à jour le contact avec ces données, saisissez \"oui\" : ");
-                }
-                if ($updValid === "oui"){
-                    $updateContact = $updateRepository->updateContact($contactId, $name, $email, $phone);
-                } else {
-                    echo "Vous pouvez recommencer votre modification en saisissant la commande \"modify\".\n";
+                    echo "Les données saisies ne sont pas conformes, veuillez refaire votre saisie.\n";
                 }
             }
         } else {
@@ -106,20 +125,27 @@ class Command
     public function delete($line): void
     {
         $idDeleteContact = (trim(substr($line,7)));
+        $contactId = $idDeleteContact;
         if (is_numeric($idDeleteContact)){
-            $Contact = new Command();
-            $Contact->detail($idDeleteContact);    
-            $delValid = readline("Cette action est irréversible, êtes-vous sûr ?");
-        } else {
-            echo "Vous avez saisi : $idDeleteContact\n";
-            $idDeleteContact = Readline("Pour supprimer un contact, vous devez saisir un Numéro valide :");
-        }
-        if ($delValid === "oui"){
-            $lineVerif = preg_match('/delete/', $line, $matches);
-            $idDeleteContact = trim(str_replace($matches, '',$line));
             $deleteRepository = new ContactManager();
             $deleteRepository->connection = new DBConnect();
-            $deleteContact = $deleteRepository->deleteContact($idDeleteContact);
+            $deleteContactName = $deleteRepository->findById($contactId);
+            if ($deleteContactName->getName() === null){
+                echo "Aucun contact avec cet identifiant, saisissez la commande \"list\" pour voir les identifiants disponibles.\n"; 
+            } else {
+                $Contact = new Command();
+                $Contact->detail($idDeleteContact);      
+                $delValid = readline("Cette action est irréversible, êtes-vous sûr ?");
+                if ($delValid === "oui"){
+                    $lineVerif = preg_match('/delete/', $line, $matches);
+                    $idDeleteContact = trim(str_replace($matches, '',$line));
+                    $deleteContact = $deleteRepository->deleteContact($idDeleteContact);
+                } else {
+                    echo "Action annulée, aucun contact n'a été supprimé.\n";
+                }
+            }        
+        } else {
+            echo "L'identifiant doit être un chiffre seul.\n";
         }
     }
